backend/requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.22.0
sqlalchemy==1.4.46
asyncpg==0.27.0
PyJWT==2.6.0
passlib[bcrypt]==1.7.4
bcrypt>=3.2.0
python-multipart==0.0.5
python-dotenv==0.21.0
